refactor Review and Aggregate ScoreIndicator style
